requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.3.1
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
